AIS-1447 - changes for application resources API
