+++ ai-services/internal/pkg/catalog/apiserver/repository/application_service/openshift.go
@@ -7,6 +7,7 @@ import (
 	"github.com/google/uuid"
 	apimodels "github.com/project-ai-services/ai-services/internal/pkg/catalog/apiserver/models"
 	"github.com/project-ai-services/ai-services/internal/pkg/catalog/types"
+	catalogutils "github.com/project-ai-services/ai-services/internal/pkg/catalog/utils"
 	runtimeTypes "github.com/project-ai-services/ai-services/internal/pkg/runtime/types"
 )
 
@@ -28,8 +29,11 @@ func (s *OpenShiftApplicationService) CreateApplication(ctx context.Context, req
 	return s.ApplicationServiceBase.CreateApplication(ctx, req, runtimeTypes.RuntimeTypeOpenShift)
 }
 
-func (s *OpenShiftApplicationService) GetApplicationResources(_ context.Context, _ uuid.UUID) (*types.ApplicationResourcesResponse, error) {
-	return nil, errOpenShiftNotSupported
+// GetApplicationResources retrieves CPU, memory, and Spyre-card usage for an application
+// using the OpenShift runtime. Each application is deployed into its own namespace
+// (ai-services-<first 8 chars of UUID>), so the runtime client is created with that namespace.
+func (s *OpenShiftApplicationService) GetApplicationResources(ctx context.Context, id uuid.UUID) (*types.ApplicationResourcesResponse, error) {
+	return s.ApplicationServiceBase.GetApplicationResources(ctx, id, catalogutils.AppNamespace(id))
 }
 
 func (s *OpenShiftApplicationService) ApplicationsPs(_ context.Context, _ uuid.UUID) (*types.ApplicationPSResponse, error) {
